tree: extract cql_types module
There are plenty of top-level modules in the crate that contain CQL
types (collection, tuple, user_type, uuid, inet, date_time, value).
This change moves them all into a single cql_types module, which
should make it easier to find them and understand their purpose.

This is a purely organizational change, no logic was changed.
