+++ scylla-rust-wrapper/src/batch.rs
@@ -6,11 +6,11 @@ use crate::cass_error::CassError;
 pub use crate::cass_types::CassBatchType;
 use crate::cass_types::{CassConsistency, make_batch_type};
 use crate::config_value::{MaybeUnsetConfig, RequestTimeout};
+use crate::cql_types::value::CassCqlValue;
 use crate::exec_profile::PerStatementExecProfile;
 use crate::retry_policy::CassRetryPolicy;
 use crate::statement::{BoundStatement, CassStatement};
 use crate::types::*;
-use crate::value::CassCqlValue;
 use scylla::statement::batch::Batch;
 use scylla::statement::{Consistency, SerialConsistency};
 use scylla::value::MaybeUnset;
