+++ scylla-rust-wrapper/src/cluster.rs
@@ -2,6 +2,7 @@ use crate::argconv::*;
 use crate::cass_error::CassError;
 use crate::cass_types::CassConsistency;
 use crate::config_value::MaybeUnsetConfig;
+use crate::cql_types::uuid::CassUuid;
 use crate::exec_profile::{CassExecProfile, ExecProfileName, exec_profile_builder_modify};
 use crate::load_balancing::{
     CassHostFilter, DcRestriction, LoadBalancingConfig, LoadBalancingKind,
@@ -11,7 +12,6 @@ use crate::runtime::{RUNTIMES, Runtime};
 use crate::ssl::CassSsl;
 use crate::timestamp_generator::CassTimestampGen;
 use crate::types::*;
-use crate::uuid::CassUuid;
 use openssl::ssl::SslContextBuilder;
 use openssl_sys::SSL_CTX_up_ref;
 use scylla::client::execution_profile::ExecutionProfileBuilder;
