+++ scylla-rust-wrapper/src/cql_types/mod.rs
@@ -1,0 +1,7 @@
+pub(crate) mod collection;
+pub(crate) mod date_time;
+pub(crate) mod inet;
+pub(crate) mod tuple;
+pub(crate) mod user_type;
+pub(crate) mod uuid;
+pub(crate) mod value;
