+++ scylla-rust-wrapper/src/cql_types/uuid.rs
