+++ scylla-rust-wrapper/src/lib.rs
@@ -14,12 +14,10 @@ pub(crate) mod batch;
 pub(crate) mod cass_error;
 pub(crate) mod cass_types;
 pub(crate) mod cluster;
-pub(crate) mod collection;
 pub(crate) mod config_value;
-pub(crate) mod date_time;
+pub(crate) mod cql_types;
 pub(crate) mod exec_profile;
 pub(crate) mod future;
-pub(crate) mod inet;
 #[cfg(cpp_integration_testing)]
 pub(crate) mod integration_testing;
 pub(crate) mod iterator;
@@ -39,10 +37,6 @@ pub(crate) mod statement;
 #[cfg(test)]
 pub(crate) mod testing;
 pub(crate) mod timestamp_generator;
-pub(crate) mod tuple;
-pub(crate) mod user_type;
-pub(crate) mod uuid;
-pub(crate) mod value;
 
 /// Includes a file generated by bindgen called `filename`.
 macro_rules! include_bindgen_generated {
