+++ scylla-rust-wrapper/src/session.rs
@@ -4,6 +4,7 @@ use crate::cass_error::*;
 use crate::cass_metrics_types::CassMetrics;
 use crate::cass_types::get_column_type;
 use crate::cluster::CassCluster;
+use crate::cql_types::uuid::CassUuid;
 use crate::exec_profile::{CassExecProfile, ExecProfileName, PerStatementExecProfile};
 use crate::future::{CassFuture, CassFutureResult, CassResultValue};
 use crate::metadata::create_table_metadata;
@@ -13,7 +14,6 @@ use crate::query_result::{CassResult, CassResultKind, CassResultMetadata};
 use crate::runtime::Runtime;
 use crate::statement::{BoundStatement, CassStatement, SimpleQueryRowSerializer};
 use crate::types::size_t;
-use crate::uuid::CassUuid;
 use scylla::client::execution_profile::ExecutionProfileHandle;
 use scylla::client::session::Session;
 use scylla::client::session_builder::SessionBuilder;
